+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
     "object-collection",
     "timer-utils",
     "work-steal-queue",
-    "work-steal-queue-v2",
     "base-coroutine",
     "detour2",
     "libhook",
