+++ base-coroutine/Cargo.toml
@@ -12,13 +12,10 @@ libc = "0.2.137"
 once_cell = "1.13.0"
 scopeguard = "1.1.0"
 num_cpus = "1.14.0"
-concurrent-queue = "2.0.0"
-st3 = "0.4.0"
 rayon = "1.6.1"
 object-collection = { version = "0.0.4", path = "../object-collection" }
 timer-utils = { version = "0.0.3", path = "../timer-utils" }
-work-steal-queue = { version = "0.1.0", path = "../work-steal-queue" }
-work-steal-queue-v2 = { version = "0.1.0", path = "../work-steal-queue-v2" }
+work-steal-queue = { version = "0.1.1", path = "../work-steal-queue" }
 
 [target."cfg(unix)".dependencies]
 jemalloc-sys = { version = "0.5.2+5.3.0-patched", default-features = false }
