+++ base-coroutine/src/lib.rs
@@ -14,7 +14,7 @@ pub use event_loop::interest::*;
 pub use event_loop::*;
 pub use scheduler::*;
 pub use stack::{Stack, StackError};
-pub use work_steal::*;
+pub use work_steal_queue::*;
 
 #[allow(dead_code)]
 mod id;
@@ -27,9 +27,6 @@ mod context;
 
 #[allow(dead_code)]
 pub mod coroutine;
-
-#[allow(dead_code)]
-mod work_steal;
 
 #[allow(dead_code)]
 pub mod scheduler;
