--- work-steal-queue-v2/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "work-steal-queue-v2"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-num_cpus = "1.14.0"
-st3 = "0.4.1"
-crossbeam-deque = "0.8.2"
