--- work-steal-queue-v2/src/lib.rs
@@ -1,6 +1,0 @@
-pub use rand::*;
-pub use work_steal::*;
-
-pub mod rand;
-
-pub mod work_steal;
